added support for bundler optimization
To only run bundler if really necessary the follow hack is used:
* Gemfile and Gemfile.lock are added to the docker archive in their respective
  latest revision.
* Bundler is run on these files.
* Only afterwards the actual source is extracted.

This way the bundling is only run if necessary.

The dockerbuild.Build struct now has a RubyProject flag, that will add the
required files for the locking.

The git helper library was modified to support adding files in specific
revisions (setting the modification time accordingly).