added support for map type in options
The command struct may now have a map as type. The map must have key
strings and values of either string or int64 type.
For an option -f the values could be given as
    -f.a A -f.b B
This would add keys "a" and "b" with respective values "A" and "B" to
the `-f` option's map.