cli/logger: replace with explicit print to stderr
Trying to test a CLI application (want to verify default handling works
properly) and failed to do so, due to the logger. I couldn't fetch the
output. This is because the logger is created on test run init (it's an
internal global), but changing the stdout file-descriptor happens during
testing.

Additionally help output is now sent to stderr (instead of stdout). This
is where help text should reside.